Refactor: wrap SafePty into own module
