+++ lib/fastlane_core.rb
@@ -1,7 +1,7 @@
 require 'json'
-require 'safe_pty'
 require 'fastlane_core/version'
 require 'fastlane_core/helper'
+require 'fastlane_core/runner'
 require 'fastlane_core/xcodebuild_list_output_parser'
 require 'fastlane_core/configuration/configuration'
 require 'fastlane_core/update_checker/update_checker'
